// frontend/src/types/supportRequestType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type SupportRequestType = typeof SupportRequestType[keyof typeof SupportRequestType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportRequestType = {
  BUG_REPORT: 'BUG_REPORT',
  FEATURE_REQUEST: 'FEATURE_REQUEST',
  TECHNICAL_SUPPORT: 'TECHNICAL_SUPPORT',
  COMPLAINT: 'COMPLAINT',
  OTHER: 'OTHER',
} as const;