// frontend/src/types/supportResponseStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type SupportResponseStatus = typeof SupportResponseStatus[keyof typeof SupportResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupportResponseStatus = {
  OPEN: 'OPEN',
  RESOLVED: 'RESOLVED',
  CLOSED: 'CLOSED',
} as const;